Add files via upload
This is a code that can be used to build a WhatsApp bot using Node.js and OpenAI. The bot is designed to understand natural language and can provide information on various topics, answer questions, give advice, and engage in conversations with users. By using this source code, you can create a powerful bot that serves multiple purposes. It is important to avoid making changes to the index.js file. 

To install, clone the repository, navigate to the downloaded folder, install the necessary packages, and start the bot using the command 'node index.js'. The API keys need to be set up in the accesser.json file.
 
To obtain the OpenAI API, visit Visit: https://beta.openai.com/account/api-keys